// frontend/lib/api/generated/model/pageContratoListOut.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * SIGA-UNMSM API
 * API del Sistema Integral de Gestión de Almacén y Abastecimiento Alimentario — UNMSM. Módulos activos: Autenticación, Usuarios, Almacenes, Catálogo Nutricional.
 * OpenAPI spec version: 0.1.0
 */
import type { ContratoListOut } from './contratoListOut';

export interface PageContratoListOut {
  items: ContratoListOut[];
  total: number;
  page: number;
  page_size: number;
}